Split styleguide fontsizes for android specific font sizes
