+++ packages/styleguide/fonts/font-sizes-base.js
@@ -1,0 +1,28 @@
+const fontSizes = {
+  articleHeadline: 45,
+  articleMeta: 15,
+  body: 18,
+  bodyMobile: 17,
+  caption: 13,
+  cardHeadline: 27,
+  cardMeta: 13,
+  cardMetaMobile: 12,
+  credits: 9,
+  flags: 12,
+  headline: 30,
+  leadHeadline: 35,
+  link: 13,
+  meta: 14,
+  pageComponentHeadline: 25,
+  pageHeadline: 40,
+  pagingMeta: 15,
+  puffLink: 11,
+  secondary: 16,
+  sliceHeadline: 32,
+  smallestHeadline: 20,
+  smallHeadline: 22,
+  teaser: 14,
+  tertiary: 15
+};
+
+export default fontSizes;
