+++ packages/styleguide/fonts/font-sizes.android.js
@@ -1,0 +1,8 @@
+import fontSizes from "./font-sizes";
+
+const androidFontSizes = {
+  ...fontSizes,
+  flags: 14
+};
+
+export default androidFontSizes;
